Add listview.js in module
