+++ index.html
@@ -12,9 +12,9 @@
         <nav class="container">
             <h2>Awesome Books</h2>
             <ul>
-                <li><a id="listBtn" href="#bookList">List</a> </li>
-                <li><a id="addBtn" href="#bookForm">Add new</a></li>
-                <li><a id="contactBtn" href="#contactDetails">Contact</a></li>
+                <li><a id="listBtn" class="change" href="#bookList">List</a> </li>
+                <li><a id="addBtn" class="change" href="#bookForm">Add new</a></li>
+                <li><a id="contactBtn" class="change" href="#contactDetails">Contact</a></li>
             </ul>
         </nav>
     </header>
