+++ index.js
@@ -1,22 +1,11 @@
 import displayBook from './modules/displayBook.js';
 import { DateTime } from './modules/luxon.js';
+import listView from './modules/ListView.js';
 
 const bookName = document.querySelector('.bookName');
 const bookAuthor = document.querySelector('.bookAuthor');
 const form = document.querySelector('.addBook');
 const bookList = JSON.parse(localStorage.getItem('bookList')) || [];
-
-const bookForm = document.querySelector('#bookForm');
-const bookSection = document.querySelector('#bookList');
-const contactSection = document.querySelector('#contactDetails');
-
-const listBtn = document.getElementById('listBtn');
-const formBtn = document.querySelector('#addBtn');
-const contactBtn = document.getElementById('contactBtn');
-
-const listArr = [bookForm, contactSection];
-const formArr = [bookSection, contactSection];
-const contactArr = [bookForm, bookSection];
 
 const message = document.querySelector('.message');
 
@@ -35,10 +24,7 @@ class Book {
 }
 
 function redirect() {
-  bookSection.classList.remove('hide');
-  listArr.forEach((item) => {
-    item.classList.add('hide');
-  });
+  listView.changeContent('List');
 }
 
 form.addEventListener('submit', (e) => {
@@ -60,46 +46,13 @@ if (!bookList.length) {
   message.innerHTML = 'click Add new link to add a new book';
 }
 
-listBtn.addEventListener('click', () => {
-  listBtn.style.color = '#e90074';
-  formBtn.style.color = '#555';
-  contactBtn.style.color = '#555';
-  bookSection.classList.remove('hide');
-  listArr.forEach((item) => {
-    item.classList.add('hide');
-  });
-});
-
-formBtn.addEventListener('click', () => {
-  listBtn.style.color = '#555';
-  formBtn.style.color = '#e90074';
-  contactBtn.style.color = '#555';
-  bookForm.classList.remove('hide');
-  formArr.forEach((item) => {
-    item.classList.add('hide');
-  });
-});
-
-contactBtn.addEventListener('click', () => {
-  listBtn.style.color = '#555';
-  formBtn.style.color = '#555';
-  contactBtn.style.color = '#e90074';
-  contactSection.classList.remove('hide');
-  contactArr.forEach((item) => {
-    item.classList.add('hide');
-  });
-});
-
-const initial = () => {
-  bookSection.classList.remove('hide');
-  listArr.forEach((item) => {
-    item.classList.add('hide');
-  });
-};
-
 setInterval(() => {
   const now = DateTime.now();
   time.textContent = now.toLocaleString(DateTime.DATETIME_MED);
 }, 1000);
 
-window.onload = initial();
+document.querySelectorAll('.change').forEach((item) => {
+  item.addEventListener('click', () => {
+    listView.changeContent(item.textContent);
+  });
+});
